Correct where temporary per-device JSON was stored.
Also, don't delete the work directory when we are debugging.

+++ spoon/src/main/java/com/squareup/spoon/Spoon.java
@@ -128,10 +128,12 @@ public final class Spoon {
       }
     }
 
-    // Clean up anything in the work directory.
-    try {
-      FileUtils.deleteDirectory(new File(output, SpoonDeviceRunner.TEMP_DIR));
-    } catch (IOException ignored) {
+    if (!debug) {
+      // Clean up anything in the work directory.
+      try {
+        FileUtils.deleteDirectory(new File(output, SpoonDeviceRunner.TEMP_DIR));
+      } catch (IOException ignored) {
+      }
     }
 
     return summary.end().build();
